Ported changes from ZDoom revision 1018-1024 / GZDoom revision 117: - While doing the interpolation rewrite I noticed that DScroller and DPolyAction were doing some things in their destructor that needed to be done in the Destroy method. - Rewrote the interpolation code. Interpolations are no longer some objects that are separate from the rest of the engine. Instead, they are owned by the thinkers starting them. Also, polyobjects only spawn a single interpolation for each polyobject instead of a single one for each vertex. Also, different types of interpolation objects are used for different types of interpolation so that they can do some additional work if eventually needed. - Fixed: The EndSequence structure was not fully initialized. - bumped min. savegame version to r1018. - Removed the 8-character limit on endpic names from the parser. (Though it might still be present in the texture manager; I don't remember.) - Fixed: EndSequence needs a proper constructor. - Some more GCC warning removals. - Repositioned the declaration of the file string in D_DoomMain() so that it won't be left on the stack at exit. - Fixed: PSymbol needs a virtual destructor so that PSymbolActionFunction can free its Arguments. - Symbols for native classes are now freed on exit. - Fixed: When following links in S_StartSound(), the function used sfx instead of S_Sfx as the array base for getting the NearLimit. - Fixed: There was still an inventory bar check in FMugShot::GetFace() after separating it from DDoomStatusBar.
SVN r1409 (latestzdoom)

+++ tools/makewad/makewad.c
@@ -148,7 +148,7 @@ int appendtozip (zipFile zipfile, const char * zipname, const char *filename)
 	if (readbuf == NULL)
 	{
 		fclose(lumpfile);
-		fprintf (stderr, "Could not allocate %d bytes\n", len);
+		fprintf (stderr, "Could not allocate %u bytes\n", (int)len);
 		return 1;
 	}
 	// read the whole source file into buffer
@@ -323,7 +323,7 @@ int buildwad (FILE *listfile, char *listfilename, const char *makecmd, char *mak
 				{
 					fprintf (wadfile, " \\\n\t\"%s\"", filename);
 				}
-				// ARGH! Stupid make. Too bad but then these files can't be checked. 
+				// ARGH! Stupid make. Too bad, but then these files can't be checked. 
 				// Still better than something that doesn't work at all though!
 				else if (!strpbrk(filename, "[]^"))	
 				{
